ipaldap: Override conversion of nsds5replicalastupdatestart and nsds5replicalastupdateend
The replication related attributes nsds5replicalastupdatestart and
nsds5replicalastupdateend have special behaviour implemented in 389,
as follows:

In case they are explicitly requested for and not set, 0 is returned.

However, 0 is not a valid value for LDAP Generalized time. Thus
we need to add these attributes to the _SYNTAX_OVERRIDE dictionary,
overriding their conversion to datetime and converting them to
string instead, which perserves the old behaviour expected by the
replication codebase.

https://fedorahosted.org/freeipa/ticket/4350